requirements.txt
fastapi==0.66.0
pydantic==1.8.2
uvicorn==0.14.0
PyJWT==2.0.0